Updated openstack/openstack Project: openstack/horizon  a09b5fe664a8cd5817ba9b6c8f68f256ea2e4d37
Cleaned up dict.update(single elem dict) code

There's several places in the code where this type of thing exists:
    form_data.update({'vpnservice_id': vpnservice.id})
better would be this:
    form_data['vpnservice_id'] = vpnservice.id
This type of thing bothers me, so I'm going to clean it up.

Also fixed the two cases of:
    if <some_string>.find(<some_other_string>) != -1:
which should clearly use the "in" operator instead.

Change-Id: Ic72a5b0a17c211f3949edc69732c1101c24590ec
Closes-Bug: 1339849