Updated openstack/openstack Project: openstack/ceilometer  33d3a9682206e1bd92e5437274873711a7d14dbb
Fix incorrect use of timestamp in test

The timestamp used has a very slim chance of being in
a different second than actually desired which would
tickle the bug. The bug itself is simply that the
mapping between sample and timestamp was incorrect.

To make sure that there isn't confusion over distinct
timestamps, _make_timestamps has been changed so that
if multiple timestamps are requested, each is a second
after the previous.

Change-Id: I22d999c506a30d4ca6b8dd7e9de501570257b99e
Closes-bug: 1341142