Updated openstack/openstack Project: openstack/trove  8b35dacab526da8deac9849433b63e58a729320a
Ensure routing key is specified in the address for a direct producer

This change is already merged in oslo-incubator.  Note that this is
not a clean merge of the fix, as Trove's copy of rpc from
oslo-incubator is *VERY* far out of date.  However, the bug still
existed in the code.  The fix just looks slightly different.

Original commit message body includes:

    Porting this fix from oslo.messaging.  This fixes the impl_qpid.py
    driver to allow it to work with the latest stable upstream QPID broker
    (version 0.28).  See the Apache Qpid Jira bug
    https://issues.apache.org/jira/browse/QPID-5557

Change-Id: If71f78e50f8a9b3acfd1e9d02c8271f17c4ebee7
Related-Bug: #1300318