Updated openstack/openstack Project: openstack/nova  d6c2a70f79fd03ef624feddee359ee1cab50d039
Deallocate the network if rescheduling for Ironic

Ironic (and any other driver that limits the allowed MAC addresses
on an instance will fail if the rescheduled instance runs on a node
that has a different constraint for MAC addresses. To deal with this
we deallocate the network (rather than e.g. trying to lazy update it)
because Nova can't know what the implications of leaving the network
allocated are *when MAC limits are in place* - and we expect them to
be in place when external systems are constrained, so being
conservative is a good idea.

However the code to trigger this looked at dhcp options (which drivers
can safely update pre-boot) not MAC addresses (which can cause conflicts
in Neutron if left in-place).

Change-Id: I59e748db7a943d5a36d75ef63b2a1ec458c58937
Closes-Bug: #1342919