Updated openstack/openstack Project: openstack/tempest  4871ae42c2cb5046ca5bdf9e95c9aa9bd6c68704
Exclude volume tags while listing instance tags

When a volume is attached to an instance and then detached, a metadata
gets associated with the volume: 'readonly' = 'False', even if the volume
didn't have this entry before. The patch
https://review.openstack.org/#/c/70085/ against Nova adds support for
volume tags, and as a result, it breaks a test case where we are
asserting that no tags are left after deletion of them.
So this patch makes sure that we assert only for 'instance' tags, and
not volume tags.

Change-Id: Icdd8c3a2e6c9d3264afc141ad516ee06c6127336