Updated openstack/openstack Project: openstack/horizon  b04f020fc1dce27b47a174259ea3ba542127401b
'Create Volume Snapshot' form should show the quota for snapshot

When creating a new volume snapshot the window shows the quota of number
of volumes and what should be shown is the quota of snapshots.
Basically, the code was pasted from "Create Volume" and nothing has been
changed to respect the equivalent quotas.

Change-Id: Ifbfa84a008d7b61e4b53f885793c5a1750dade3c
Closes-bug: 1282172