Updated openstack/openstack Project: openstack/tempest  6ac2838b92eddea01f0bc6108dc7a19756a27a16
Skip baremetal tests if driver not supported

Adds a list of supported Ironic drivers that can be used for its API tests and
skips if the configured driver is not supported..  The list of supported
drivers will be extended/removed as we support API testing against clouds with
real drivers enabled.

Change-Id: I0b975548ff3656e8a052c79c7b9ffa1d8d721024
Partial-bug: #1334423.