Updated openstack/openstack Project: openstack/heat  9adf703a8a39e921a5a7c7158128069f223153dc
Store clients in context

Having the Stack create the Clients object means that custom constraints
do not have access to the already created clients. This greatly
complicates test mocking, and results in unnecessarily created
constraint-scoped clients.

This change makes the Clients be created in RequestContext on demand.
Constraints use the Clients associated with the context.

Mocking client create calls is simplified, generally needing to
call _create once instead of MultipleTimes()

Change-Id: If22f3258aa9ff906bd60977cc9c2752f18c11c65