Updated openstack/openstack Project: openstack/requirements  5ab09ed27c34c4fa003917baec3bf366a7ce1f26
Revert oslo.messaging 1.4.0.0a3 and oslo.config 1.4.0.0a2

This reverts commit 7916d6cb041960750604e91dd768f03700ff845d and
commit 7c6ea39940380738b4b969ed4c1b11e45ecd2055.

The oslo alpha releases are only published as wheels because we
don't want to expose them to icehouse users who may still be using
pip 1.3 which doesn't know to ignore alphas.

We recently switched to bandersnatch based pypi mirror to allow
us to mirror these wheels.

However, the py26 unit test jobs on centos6 are failing because
an old version of virtualenv and pip is installed. The failure looks
like:

  Could not find a version that satisfies the requirement oslo.config>=1.4.0.0a2

Monty has been working on a fix, but it's proving to be a whack-a-mole
situation with each fix showing up yet another issue. Let's remove
the alphas for now so that the auto-proposing bot isn't proposing
changes which we know are going to fail.

Change-Id: I52c695086b90a94eab2ab92533665b44f00b9077