Updated openstack/openstack Project: openstack/tempest  8e311a5328fbd24eebdc38e8e461b726fae555ec
Replace the console output on Error with fault a message

Trying to get the console log from an ERROR-ed server
does not really works

The console log does not contains any meaningful info
related to most status changes,
it is just what the instance prints to the serial console.

Only one action is guest assisted: the 'soft reboot',
which automatically switched to 'hard reboot'
when the guest forgets to assist
(ie. does not shutdown in 120 sec)

Testing the console output presence before reporting
active state would make more sense.

Instead of attempting to get the serial console output,
after this change the response 'fault' field will be used
for exception detail when available.

Closes-Bug: #1332414

This reverts commit c1a3bbe37c6202562fad5ff1be6e818d12797aee.

Change-Id: I2018f98cdc4a1bc7dd8f34833df228dc56d88a95