Updated openstack/openstack Project: openstack/python-neutronclient  03be2b4bda9087dc836d3a734ab830caef2dc43f
Add CLI Support for DVR

This patch modifies the existing python-neutron
client to create and list the distributed
virutal routers.
Based on the user role (admin) will be able to
create a distributed virtual router and should
be able to see the distributed nature of the
router.

Partially-implements: blueprint neutron-ovs-dvr

DocImpact

Change-Id: I96269167525563bdd8a52ff981ee5098f5d68e54