Updated openstack/openstack Project: openstack/trove  98083f6e9a5367c4b28d8dd0184d478f7bf8a750
Enhance trove-manage help

The automated tools for CLI reference document creation have been
enhanced to support *-manage utilities. Since these tools run the CLI
and extract the help from it, they are only as good as the help of the
utility in question.

The help for trove-manage has been enhanced so that the CLI reference
documentation will have useful information in it.

Change-Id: I2560e668cf52bb81b9561abc7747db2155921721
Closes-Bug: #1339840