Updated openstack/openstack Project: openstack/glance  a28db383fc95fb65b9331f58366f69d53c7ca303
Raising 400 Bad Request when using "changes-since" filter on v2

In Glance API V2, currently, the user gets a 500 error when tries
to use the "changes-since" filter.This filter, which was available
on API v1 is no longer available on V2.

This patch returns to the user a 400 Bad Request with a proper message,
instead of a generic 500 Internal Server Error.

Change-Id: Icece859a7f12f6f4aa1f7bb945134a935048b315
Closes-bug: 1273975