Updated openstack/openstack Project: openstack/cinder  e341e16a2e3b257614e1d0964c4bb5e6303e2193
Add some log info for NoValidHost

Currently when a create call fails due to a
misconfiguration on the backend, or a problem
with satisfying the extra-specs from the filter
scheduler, the only thing you get in the logs is
a "No valid host".

This patch just adds a tiny bit more info to give us
a better idea where we're at.

First the exception shouldn't be setup with "reason=""",
so at least add "reason=no weighed hosts found".

Second, add a warning message prior to getting to that
point when we fail to find a host that satisfies the
{extra/qos}-specs that were requested, and include what
was being requested.

Change-Id: I2ce65b973196f673bac455613a6f8626320339ba