Updated openstack/openstack Project: openstack/horizon  1f879455f82f0f095548906e4a87e0b580350f8d
Run router dashboard unit tests by default

Currently the "router" dashboard does not have its unit tests
run by default. This was beacuse the dashboard needed to be
turned on for these tests to be run and turning on this dashboard
was not mandatory.

However, now that the dashboards are pluggable, turning on the
router dashboard tests, and more specifically the Cisco N1K profile
tests is doable and necessary. This fix aims to remove the checks
put in place to run the tests only when the dashboard is available
and instead now run all the tests mandatorily

Closes-bug: #1229819

Change-Id: I3c77a71db08ac611e5bdf9274af5a558bfda3327