Updated openstack/openstack Project: openstack/nova  4ff26f2ca0eadfa50aec54e5ba4d1c90221b7ecf
VMware: DatastorePath join() and __eq__()

There are quite a number of use cases where we construct additional
datastore paths out of one representing some ancestor directory,
so implemented a join() method that allows for just that.
It is useful to define equality operator when comparing two DatstorePath
objects, so added that as well.

partial blueprint vmware-spawn-refactor
Change-Id: Iec44fdd0beba232b57bcd0a07c39e628208ca9b2