Updated openstack/openstack Project: openstack-dev/pbr  c2aab3b27f03887606b49711a7db9a21b72a60c1
Stop filtering out argparse

Back in the day, attempting to install argparse on python 2.7 would
result in failure, as the setup.py of argparse failed hard if it was
attempted. That seems to be fixed, and when we do filter it out of
the list now, we cause problems with python 2.6 and wheels (if we make
a wheel on 2.6 and it depends on argparse, the pbr code means that
argparse doesn't get added properly)

Remove the filtering so that our wheels can be universal.

Closes-bug: #1346357
Change-Id: I49d539ac6fb8ee75f8d32fdfbd20129b13d36499