Updated openstack/openstack Project: openstack/nova  a1c40994b953d4acaba58482adb6a4ccc9916f91
Add APIv2 support to make host optional on evacuate

In the event of an unrecoverable hardware failure, support needs to
relocate an instance to another compute so it can be rebuilt.

The changes involved in this patch are:
  [*] Add a new v2 extension to determine that the host argument
      on evacuate is now optional.(Extended_evacuate_find_host)
  [*] Doc regeneration.

DocImpact: The evacuate target host is now optional.
If 'host' field is not sent in the request, the scheduler will
determine the target host.
This will include nova client changes ( on the proper commit ) to support
this new optional parameter.

Implements: blueprint find-host-and-evacuate-instance

Change-Id: Ib34fb3120263b746ad2f8fe89c14137e36a07a53
Co-Authored-By: Juan M. Olle <juan.m.olle@intel.com>
Co-Authored-By: Andres Buraschi <andres.buraschi@intel.com>
Co-Authored-By: Anuj Mathur <anujm@thoughtworks.com>
Co-Authored-By: Navneet Kumar <navneetk@thoughtworks.com>
Co-Authored-By: Claxton Correya <claxton@gmail.com>