Updated openstack/openstack Project: openstack/cinder  b960263f35af3d5b31916f2707ea4c1d755122f6
Add additional explicit imports of _ where needed

It appears that I missed some of the needed explicit
imports of _.  I have now run through all the /bin/cinder-*
functions and ensured that they can be properly started.

This patch adds additional imports of _ and also changes
cinder-rtstool to use the 'cinder' domain as was suggested by
Duncan Thomas in the first review cycle.

Change-Id: I6477093cb817b5b9506dd3787f027e073c51f37d
Closes-Bug:  1345789