Updated openstack/openstack Project: openstack/glance  81e4a01818aa8c1bc639e00f45158b34f2c4f894
Also run v2 functional tests with registry

Currently the v2 functional tests are run without the v2 registry enabled.

This means that bugs which are only seen when the v2 registry is enabled
may not be caught by the gate (eg #1308419, #1339775).

Here we enable some basic functional testing (eg image life cycle) with
the v2 registry enabled (in addition to the existing direct-to-database
mode). This should help prevent some types of bugs being introduced.

There is a plan to move some functional tests into tempest, at which
point these additional tests can be removed in favour of equivalent
tempest tests.

Change-Id: I774a277673390650874aad4334b28d1a872ea93f