Updated openstack/openstack Project: openstack/cinder  f51621c07fc088625f27d68a8008a5f60ed5e57b
Fix failure of source volume deletion in migrate_volume_completion

In VolumeManager.migrate_volume_completion(), when volume_status is not
'retyping', status_update is not set. However, status_update is always
referenced as dictionary on volume deletion path, so it fails to delete
migration source volume. This patch fixes the issue.

Change-Id: I06889dae326b9557e044f1712d9f3205bb68a953
Closes-Bug: 1344341