Updated openstack/openstack Project: openstack/cinder  226a9511dc030ef6f4eb109bf1e4788af3a4209e
Configure write cache option of tgtd iscsi driver

Cinder LVMiSCSI driver is using default value of write-cache parameter
of tgtd iscsi driver. In this setting, write I/O from guest instance is
cached on dirty cache of a host.(write-back mode)

In this case, data lost may be occurred if the host crashes before
flushing dirty cache. This may cause a lot of instances to lose
their data.

In order to avoid this issue, it is better to turn off the write cache.
(write-through mode)

This patch adds "iscsi_write_cache" parameter to configure a behavior of
write cache. The default value is "iscsi_write_cache=on".(write-back mode)

Closes-Bug: 1336568
DocImpact

Change-Id: I7a495bc6118d4254576bdf1620a04ac537b3078d
Signed-off-by: Mitsuhiro Tanino <mitsuhiro.tanino@hds.com>