Updated openstack/openstack Project: openstack/keystone  52876f3ba52496eb853cea38f4e4e5e751734a9b
Mark the 'check_vX_token' methods deprecated

The check_v2_token and check_v3_token perform the same work as
the corresponding validate_vX_token methods on the provider
api. The validate methods are used in favor of the check methods.

This patchset marks the check methods as deprecated, slated for
removal in the K cycle.

Change-Id: I717dd93ce359439af238dcf1f3349e9477e6b82b
bp: non-persistent-tokens