Updated openstack/openstack Project: openstack/horizon  a1cf7e4626b8c13c3b818601b72309e193299502
Add support for trove incremental backups

Added extra column to backups tables (under instances
and under backups) that will show Yes / No if the
backup is incremental (ie it has a parent)

When you drill down into backup details we now display
the name of the backup parent if this is an
incremental backup. This is displayed as a link with
the backup name so it is navigable

When you create a new backup you can now specify
an existing backup to use as the parent

Note: At some point we should probably have the
create backup dialog filter the available parents
based on the current selection for the instance.
However, so far I haven't worked out how to make
the backup parent choiceField change dynamically
based on user updates to the instance choiceField

Change-Id: I8bc70805aae50a019da3915f8e7aff25f6a7bbc6
Implements: blueprint trove-incremental-backup