Updated openstack/openstack Project: openstack/nova  b0f5992c5f2647cfa6026a8c07e146edc46e137c
Remove unneeded calls in test_shelve to start instances

During running of tox the system can timeout on slower systems.  This
is due to rpc calls to other threads failing in testing.  This
particular case involved shelving calls to start instances.  However,
when we started using objects, these calls became unnecessary for these
unit tests.  Thus we are removing the calls which should remove
the traceback for slow systems.

Change-Id: I52f1e86815cc2cac8dc31865f00eef2d4ffeea59
Partial-Bug: 1311778