Updated openstack/openstack Project: openstack/neutron  65477ca43c09fd2a7916ca8c328cebc53b26b111
This patch changes the name of directory from mech_arista to arista

This name change of directroy is needed so that additional drivers
can be moved into this directory and share the code among drivers.

In order to accomodate this name change, setup.cfg is modified to
point to the Arista ML2 driver at the new location. Additionally,
a similar new directroy is created for test files as well.

Change-Id: I614ccf8b6d152aa99e20ef1ea762c26115ab9373
closes-bug: 1346712